// dev_gm-smart-surveillance-3ee717e6d233/NVP/smart-surveillance/client/src/app/_models/camera.ts
export class Camera {
    _id: String;
    user_id: String;
    name: String;
    type:String;
    source: String;
    fps:String;
    frame: String;
    roi: Array<String>;
    gps: String;
    protocol: String;
    interface: String;
    status:String;
    agent:String;
    peopleCount: boolean;
    vehicleCount: boolean;
    vehicleSpeed: boolean;
    vehicleTracking: boolean;
    peopleTracking: boolean;
    peopleList: boolean;
    vehicleList:boolean;
    faceRecognition: boolean;
    intrusionDetection: boolean;
    faceAnalytics : boolean;
    deepAssurance : boolean;
    carCrashDetection : boolean;
    wrongTurnDetection : boolean;
    parkingViolation : boolean;
    crowdDetection : boolean;
    fashion : boolean;
    litterDetection : boolean;
    helmetDetection : boolean;
    climbingDetection : boolean;
    theme : boolean;
    anpr : boolean;
    violence : boolean;
    media_type: String;
    algo: Array<String>;
    people_whitelist: String;
    people_blacklist: String;
    vehicle_whitelist: String;
    vehicle_blacklist: String;
}